Decreased memory requirements for team predictor

+++ teampredictor.h
@@ -20,7 +20,7 @@ namespace technicalmachine {
 
 void predict_team (int detailed [][7], Team &team, bool using_lead = true);
 
-void predict_pokemon (Team &team, std::vector<double> estimate, int detailed [][7], double multiplier [END_SPECIES][END_SPECIES]);
+void predict_pokemon (Team &team, std::vector<float> estimate, int detailed [][7], float multiplier [END_SPECIES][END_SPECIES]);
 
 void predict_move (Pokemon &member, int detailed [][7]);
